# backend/openapi.json
{
  "components": {
    "schemas": {
      "AnswerBlock": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationRef"
            },
            "title": "Citations",
            "type": "array"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/AnswerClaim"
            },
            "title": "Claims",
            "type": "array"
          },
          "finalized": {
            "default": true,
            "title": "Finalized",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "markdown": {
            "title": "Markdown",
            "type": "string"
          }
        },
        "required": [
          "id",
          "markdown"
        ],
        "title": "AnswerBlock",
        "type": "object"
      },
...
      "ClaimEvidenceRelationView": {
        "properties": {
          "checker_method": {
            "default": "legacy",
            "title": "Checker Method",
            "type": "string"
          },
          "checker_version": {
            "default": "legacy",
            "title": "Checker Version",
            "type": "string"
          },
          "claim_id": {
            "format": "uuid",
            "title": "Claim Id",
            "type": "string"
          },
          "claim_text": {
            "title": "Claim Text",
            "type": "string"
          },
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "rationale": {
            "default": "",
            "title": "Rationale",
            "type": "string"
          },
          "relation": {
            "enum": [
              "supports",
              "contradicts",
              "contextualizes"
            ],
            "title": "Relation",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "claim_text",
          "evidence_id",
          "relation"
        ],
        "title": "ClaimEvidenceRelationView",
        "type": "object"
      },
...
      "DateWindow": {
        "properties": {
          "end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "timezone": {
            "default": "UTC",
            "title": "Timezone",
            "type": "string"
          }
        },
        "title": "DateWindow",
        "type": "object"
      },
...
      "EvidenceView": {
        "properties": {
          "asset_content_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Content Url"
          },
          "asset_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "asset_mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Mime Type"
          },
          "asset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Name"
          },
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "char_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char End"
          },
          "char_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char Start"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "document_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Version Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "locator": {
            "title": "Locator",
            "type": "string"
          },
          "locator_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locator Data"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "segment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Id"
          },
          "source": {
            "$ref": "#/components/schemas/SourceView"
          },
          "support_status": {
            "$ref": "#/components/schemas/SupportStatus"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "text",
          "locator",
          "support_status",
          "captured_at",
          "content_hash"
        ],
        "title": "EvidenceView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunCreate": {
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "date_window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DateWindow"
              },
              {
                "type": "null"
              }
            ]
          },
          "document_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 20,
            "title": "Document Ids",
            "type": "array"
          },
          "mode": {
            "$ref": "#/components/schemas/RunMode",
            "default": "quick"
          },
          "query": {
            "maxLength": 8000,
            "minLength": 2,
            "title": "Query",
            "type": "string"
          },
          "source_scope": {
            "items": {
              "enum": [
                "web",
                "academic",
                "software",
                "documents"
              ],
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Scope",
            "type": "array"
          }
        },
        "required": [
          "conversation_id",
          "query"
        ],
        "title": "RunCreate",
        "type": "object"
      },
...
      "RunMode": {
        "enum": [
          "quick",
          "research"
        ],
        "title": "RunMode",
        "type": "string"
      },
      "RunQualityView": {
        "properties": {
          "citation_resolution_rate": {
            "title": "Citation Resolution Rate",
            "type": "number"
          },
          "claim_citation_rate": {
            "title": "Claim Citation Rate",
            "type": "number"
          },
          "claim_count": {
            "title": "Claim Count",
            "type": "integer"
          },
          "distinct_source_group_count": {
            "title": "Distinct Source Group Count",
            "type": "integer"
          },
          "duplicate_sources_removed": {
            "default": 0,
            "title": "Duplicate Sources Removed",
            "type": "integer"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "evidence_relation_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Evidence Relation Counts",
            "type": "object"
          },
          "evidence_relations": {
            "items": {
              "$ref": "#/components/schemas/ClaimEvidenceRelationView"
            },
            "title": "Evidence Relations",
            "type": "array"
          },
          "facets": {
            "items": {
              "$ref": "#/components/schemas/RunFacetView"
            },
            "title": "Facets",
            "type": "array"
          },
          "gaps_count": {
            "default": 0,
            "title": "Gaps Count",
            "type": "integer"
          },
          "provider_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Provider Counts",
            "type": "object"
          },
          "queue_wait_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Wait Ms"
          },
          "risky_source_events": {
            "default": 0,
            "title": "Risky Source Events",
            "type": "integer"
          },
          "run_elapsed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Elapsed Ms"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "source_count": {
            "title": "Source Count",
            "type": "integer"
          },
          "source_kind_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Source Kind Counts",
            "type": "object"
          },
          "stage_timings_ms": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Stage Timings Ms",
            "type": "object"
          },
          "support_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Support Counts",
            "type": "object"
          },
          "supported_claim_rate": {
            "title": "Supported Claim Rate",
            "type": "number"
          }
        },
        "required": [
          "run_id",
          "claim_count",
          "evidence_count",
          "source_count",
          "distinct_source_group_count",
          "citation_resolution_rate",
          "claim_citation_rate",
          "supported_claim_rate"
        ],
        "title": "RunQualityView",
        "type": "object"
      },
      "RunSnapshot": {
        "properties": {
          "answer_blocks": {
            "items": {
              "$ref": "#/components/schemas/AnswerBlock"
            },
            "title": "Answer Blocks",
            "type": "array"
          },
          "budget_version": {
            "default": "legacy",
            "title": "Budget Version",
            "type": "string"
          },
          "cancellation_requested": {
            "default": false,
            "title": "Cancellation Requested",
            "type": "boolean"
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "date_window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DateWindow"
              },
              {
                "type": "null"
              }
            ]
          },
          "deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "document_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Document Ids",
            "type": "array"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "title": "Gaps",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seq": {
            "default": 0,
            "title": "Last Seq",
            "type": "integer"
          },
          "mode": {
            "$ref": "#/components/schemas/RunMode"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "source_scope": {
            "items": {
              "type": "string"
            },
            "title": "Source Scope",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "usage_ledger": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Usage Ledger",
            "type": "object"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "query",
          "mode",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "RunSnapshot",
        "type": "object"
      },
      "RunStatus": {
        "enum": [
          "queued",
          "planning",
          "discovering",
          "reading",
          "extracting",
          "checking",
          "synthesizing",
          "completed",
          "partial",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "SupportStatus": {
        "enum": [
          "supported",
          "partially_supported",
          "conflicting",
          "insufficient_evidence"
        ],
        "title": "SupportStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VisualizationView": {
        "properties": {
          "approved_spec": {
            "$ref": "#/components/schemas/VisualizationSpec"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data_lineage": {
            "items": {
              "$ref": "#/components/schemas/VisualizationLineageRef"
            },
            "title": "Data Lineage",
            "type": "array"
          },
          "dataset": {
            "additionalProperties": true,
            "title": "Dataset",
            "type": "object"
          },
          "dataset_id": {
            "format": "uuid",
            "title": "Dataset Id",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "export_metadata": {
            "additionalProperties": true,
            "title": "Export Metadata",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/VisualizationKind"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "dataset_id",
          "kind",
          "title",
          "dataset",
          "approved_spec",
          "created_at"
        ],
        "title": "VisualizationView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ARES API",
    "version": "0.11.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs": {
      "post": {
        "operationId": "create_run_api_v1_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run"
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run"
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_api_v1_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run"
      }
    },
...
    "/api/v1/runs/{run_id}/evidence": {
      "get": {
        "operationId": "list_run_evidence_api_v1_runs__run_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvidenceView"
                  },
                  "title": "Response List Run Evidence Api V1 Runs  Run Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Run Evidence"
      }
    },
...
    "/api/v1/runs/{run_id}/quality": {
      "get": {
        "operationId": "get_run_quality_api_v1_runs__run_id__quality_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunQualityView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Quality"
      }
    },
...
    "/api/v2/runs/{run_id}/visualizations": {
      "get": {
        "operationId": "get_run_visualizations_api_v2_runs__run_id__visualizations_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VisualizationView"
                  },
                  "title": "Response Get Run Visualizations Api V2 Runs  Run Id  Visualizations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Visualizations"
      }
    },
    "/api/v2/runs/{run_id}/visualizations/{visualization_id}/export.csv": {
      "get": {
        "operationId": "export_run_visualization_csv_api_v2_runs__run_id__visualizations__visualization_id__export_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "visualization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visualization Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Lineage-bearing CSV export"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Run Visualization Csv"
      }
    },
...
  }
}